+++ errai-common/src/main/java/org/jboss/errai/common/metadata/DeploymentContext.java
@@ -40,12 +40,12 @@ import org.slf4j.LoggerFactory;
  * @author Heiko Braun <hbraun@redhat.com>
  */
 public class DeploymentContext {
-  private List<URL> configUrls;
-  private Map<String, File> subContexts = new HashMap<String, File>();
-  private Set<String> processedUrls = new HashSet<String>();
-  private Set<File> createdTmpFiles = new HashSet<File>();
+  private final List<URL> configUrls;
+  private final Map<String, File> subContexts = new HashMap<String, File>();
+  private final Set<String> processedUrls = new HashSet<String>();
+  private final Set<File> createdTmpFiles = new HashSet<File>();
 
-  private Logger log = LoggerFactory.getLogger(DeploymentContext.class);
+  private final Logger log = LoggerFactory.getLogger(DeploymentContext.class);
 
   public DeploymentContext(List<URL> configUrls) {
     this.configUrls = configUrls;
@@ -72,12 +72,12 @@ public class DeploymentContext {
 
     final List<URL> superAndSubContexts = new ArrayList<URL>();
 
-    for (Map.Entry<String, File> entry : subContexts.entrySet()) {
-      File unzipped = entry.getValue();
+    for (final Map.Entry<String, File> entry : subContexts.entrySet()) {
+      final File unzipped = entry.getValue();
       try {
         superAndSubContexts.add(unzipped.toURI().toURL());
       }
-      catch (MalformedURLException e) {
+      catch (final MalformedURLException e) {
         throw new RuntimeException(e);
       }
     }
@@ -92,8 +92,8 @@ public class DeploymentContext {
   }
 
   public void close() {
-    for (File f : createdTmpFiles) {
-      boolean deleted = deleteDirectory(f);
+    for (final File f : createdTmpFiles) {
+      final boolean deleted = deleteDirectory(f);
       if (!deleted) {
         //note: use an error message instead of an exception
         log.error("failed to cleanup: files were not deleted: " + f.getPath() + " (exists:" + f.exists() + ")");
@@ -103,13 +103,15 @@ public class DeploymentContext {
 
   static public boolean deleteDirectory(File path) {
     if (path.exists()) {
-      File[] files = path.listFiles();
-      for (int i = 0; i < files.length; i++) {
-        if (files[i].isDirectory()) {
-          deleteDirectory(files[i]);
-        }
-        else {
-          files[i].delete();
+      final File[] files = path.listFiles();
+      if (files != null) {
+        for (int i = 0; i < files.length; i++) {
+          if (files[i].isDirectory()) {
+            deleteDirectory(files[i]);
+          }
+          else {
+            files[i].delete();
+          }
         }
       }
     }
